Rename folders and add empty readme files by practice.
