+++ practices/1 - first steps with flex/exercise_1.l
