+++ practices/1 - first steps with flex/exercise_2.l
