+++ practices/1 - first steps with flex/exercise_3.l
