+++ practices/1 - first steps with flex/exercise_4.l
