+++ practices/1 - first steps with flex/exercise_5.l
