+++ practices/1 - first steps with flex/exercise_6.l
