+++ practices/1 - first steps with flex/exercise_7.l
