+++ practices/1 - first steps with flex/input_1.txt
