+++ practices/1 - first steps with flex/input_4.txt
