+++ practices/5 - linked list/compilador.y
