+++ practices/5 - linked list/compilador_tabla_arreglos.y
