+++ practices/5 - linked list/ejemplo.txt
